requirements.txt
flask
flask-cors
pandas
numpy
tensorflow
joblib
google-generativeai
